requirements.txt
attrs
autopep8
cattrs
certifi
charset-normalizer
chess
Cython
fastjsonschema
idna
numpy
pycodestyle
python-chess
requests
typing_extensions
urllib3
zstandard
opencv-python
torch
torchvision
pandas
tqdm
matplotlib
Pillow
